backend/requirements.txt
fastapi==0.95.0
pydantic==1.10.9
numpy==1.24.2
pandas==2.0.0
uvicorn==0.21.1
peewee==3.16.2
requests==2.30.0
stem==1.8.1
typing==3.7.4.3
fastapi-cache2==0.2.1
websockets==11.0.3